+++ src/main/java/com/example/shimpyo/domain/auth/dto/MailCodeSendDto.java
@@ -1,0 +1,14 @@
+package com.example.shimpyo.domain.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MailCodeSendDto {
+
+    private String email;
+    private String type;
+}
