+++ src/main/java/com/example/shimpyo/domain/auth/dto/RegisterUserRequest.java
@@ -1,8 +1,8 @@
-package com.example.shimpyo.domain.user.dto;
+package com.example.shimpyo.domain.auth.dto;
 
 import com.example.shimpyo.domain.user.entity.SocialType;
 import com.example.shimpyo.domain.user.entity.User;
-import com.example.shimpyo.domain.user.entity.UserAuth;
+import com.example.shimpyo.domain.auth.entity.UserAuth;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
