+++ src/main/java/com/example/shimpyo/domain/user/controller/UserController.java
@@ -1,9 +1,5 @@
 package com.example.shimpyo.domain.user.controller;
 
-import com.example.shimpyo.domain.auth.dto.MailVerifyDto;
-import com.example.shimpyo.domain.auth.service.MailService;
-import com.example.shimpyo.domain.user.service.AuthService;
-import com.example.shimpyo.domain.user.dto.RegisterUserRequest;
 import com.example.shimpyo.domain.user.service.UserService;
 import com.example.shimpyo.global.BaseException;
 import com.example.shimpyo.global.exceptionType.MemberExceptionType;
@@ -22,43 +18,7 @@ import java.util.Map;
 @RestController
 public class UserController {
 
-    private final AuthService authService;
-    private final MailService mailService;
     private final UserService userService;
-
-    // [#MOO1] 사용자 회원가입 시작
-    @PostMapping("/signup")
-    public ResponseEntity<?> registerUser(@RequestBody RegisterUserRequest dto){
-        authService.registerUser(dto);
-        return ResponseEntity.ok("회원가입 완료");
-    }
-    // [#MOO1] 사용자 회원가입 끝
-
-    // [#M002] 이메일 검증 시작
-    @GetMapping("/check/email/duplicate")
-    public ResponseEntity<?> getEmail(@RequestParam String email){
-        return ResponseEntity.ok(authService.emailCheck(email));
-    }
-    // [#M002] 이메일 검증 끝
-
-    // [#MOO4] 이메일 인증 코드 발급 시작
-    @PostMapping("/email")
-    public ResponseEntity<?> sendEmail(@RequestParam String email){
-        mailService.authEmail(email);
-
-        return ResponseEntity.ok("이메일을 전송하였습니다.");
-    }
-    // [#MOO4] 이메일 인증 코드 발급 끝
-
-    // [#MOO5] 이메일 인증 코드 검증 시작
-    @PostMapping("/email/verify")
-    public ResponseEntity<?> verifyEmail(@RequestBody MailVerifyDto dto){
-        mailService.verifyAuthCode(dto);
-
-        return ResponseEntity.ok("인증 완료");
-    }
-    // [#MOO5] 이메일 인증 코드 검증 끝
-
 
     @PatchMapping("/mypage")
     public ResponseEntity<Void> changeNickname(@AuthenticationPrincipal UserDetails user,
