--- src/main/java/com/example/shimpyo/global/exceptionType/readMe.md
@@ -1,6 +1,0 @@
-* 사용 방법
-* import com.example.shimpyo.global.BaseException;
-* import static com.example.shimpyo.global.exceptionType.MemberExceptionType.*;
-
-
-* User user = userRepository.findById(~~~).orElseThrow(() -> new Exception(MEMBER_NOT_FOUND);
